Remove NotificationsController and most `==` methods
