+++ Persephone/App.swift
@@ -12,7 +12,7 @@ import ReSwift
 struct App {
   static let store = Store<AppState>(reducer: appReducer, state: nil)
   static let trackTimer = TrackTimer()
-  static let mpdClient = MPDClient(withDelegate: NotificationsController())
   static let userNotificationsController = UserNotificationsController()
   static let mpdServerController = MPDServerController()
+  static let mpdClient = MPDClient(withDelegate: mpdServerController)
 }
