+++ Persephone/Controllers/MPDServerController.swift
@@ -17,11 +17,15 @@ class MPDServerController {
   }
 
   func connect() {
-    App.store.dispatch(MPDConnectAction())
+    DispatchQueue.main.async {
+      App.store.dispatch(MPDConnectAction())
+    }
   }
 
   func disconnect() {
-    App.store.dispatch(MPDDisconnectAction())
+    DispatchQueue.main.async {
+      App.store.dispatch(MPDDisconnectAction())
+    }
   }
 }
 
@@ -31,5 +35,51 @@ extension MPDServerController: StoreSubscriber {
   func newState(state: MPDServer) {
     disconnect()
     connect()
+  }
+}
+
+extension MPDServerController: MPDClientDelegate {
+  func didConnect(mpdClient: MPDClient) {}
+
+  func willDisconnect(mpdClient: MPDClient) {
+    DispatchQueue.main.async {
+      App.store.dispatch(UpdateAlbumListAction(albums: []))
+    }
+  }
+
+  func didUpdateStatus(mpdClient: MPDClient, status: MPDClient.MPDStatus) {
+    DispatchQueue.main.async {
+      App.store.dispatch(UpdateStatusAction(status: status))
+    }
+  }
+
+  func willStartDatabaseUpdate(mpdClient: MPDClient) {
+    DispatchQueue.main.async {
+      App.store.dispatch(StartedDatabaseUpdateAction())
+    }
+  }
+
+  func didFinishDatabaseUpdate(mpdClient: MPDClient) {
+    DispatchQueue.main.async {
+      App.store.dispatch(FinishedDatabaseUpdateAction())
+    }
+  }
+
+  func didUpdateQueue(mpdClient: MPDClient, queue: [MPDClient.MPDSong]) {
+    DispatchQueue.main.async {
+      App.store.dispatch(UpdateQueueAction(queue: queue))
+    }
+  }
+
+  func didUpdateQueuePos(mpdClient: MPDClient, song: Int) {
+    DispatchQueue.main.async {
+      App.store.dispatch(UpdateQueuePosAction(queuePos: song))
+    }
+  }
+
+  func didLoadAlbums(mpdClient: MPDClient, albums: [MPDClient.MPDAlbum]) {
+    DispatchQueue.main.async {
+      App.store.dispatch(UpdateAlbumListAction(albums: albums))
+    }
   }
 }
