+++ Persephone/State/QueueState.swift
@@ -8,17 +8,9 @@
 
 import ReSwift
 
-struct QueueState: StateType {
+struct QueueState: StateType, Equatable {
   var queue: [QueueItem] = []
   var queuePos: Int = -1
 
   var state: MPDClient.MPDStatus.State?
-}
-
-extension QueueState: Equatable {
-  static func == (lhs: QueueState, rhs: QueueState) -> Bool {
-    return (lhs.queue == rhs.queue) &&
-      (lhs.queuePos == rhs.queuePos) &&
-      (lhs.state == rhs.state)
-  }
 }
